+++ src/lib/Components/Consignments/Screens/__tests__/SelectFromPhotoLibrary-tests.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import "react-native"
+import { Alert, Linking, NativeModules } from "react-native"
 import * as renderer from "react-test-renderer"
 
 import SelectFromPhotoLibrary from "../SelectFromPhotoLibrary"
@@ -7,6 +7,9 @@ import SelectFromPhotoLibrary from "../SelectFromPhotoLibrary"
 jest.mock("lib/NativeModules/triggerCamera", () => ({ triggerCamera: jest.fn() }))
 import { triggerCamera } from "lib/NativeModules/triggerCamera"
 const triggerMock = triggerCamera as jest.Mock<any>
+
+jest.mock("Alert", () => ({ alert: jest.fn() }))
+jest.mock("Linking", () => ({ openURL: jest.fn() }))
 
 const nav = {} as any
 const route = {} as any
@@ -38,5 +41,78 @@ it("adds new photo to the list, and selects it", () => {
       cameraImages: [{ image: { url: "https://image.com" } }],
       selection: expect.anything(),
     })
+  })
+})
+
+describe("concerning camera errors", () => {
+  let alert: jest.Mock<typeof Alert.alert> = null
+  const { ARTakeCameraPhotoModule } = NativeModules
+
+  beforeEach(() => {
+    alert = Alert.alert as any
+    alert.mockReset()
+  })
+
+  it("shows an alert when no camera is available", async () => {
+    triggerMock.mockImplementationOnce(() =>
+      Promise.reject({
+        code: ARTakeCameraPhotoModule.errorCodes.cameraNotAvailable,
+        message: "Camera not available",
+      })
+    )
+    const select = new SelectFromPhotoLibrary(emptyProps)
+    await select.onPressNewPhoto()
+    expect(alert).toHaveBeenCalledWith("Camera not available")
+  })
+
+  it("shows an alert when the camera cannot produce media of type image", async () => {
+    triggerMock.mockImplementationOnce(() =>
+      Promise.reject({
+        code: ARTakeCameraPhotoModule.errorCodes.imageMediaNotAvailable,
+        message: "Camera can’t take photos",
+      })
+    )
+    const select = new SelectFromPhotoLibrary(emptyProps)
+    await select.onPressNewPhoto()
+    expect(alert).toHaveBeenCalledWith("Camera can’t take photos")
+  })
+
+  it("shows an alert that links to Settings.app when the user has denied access to the camera", async () => {
+    triggerMock.mockImplementationOnce(() =>
+      Promise.reject({
+        code: ARTakeCameraPhotoModule.errorCodes.cameraAccessDenied,
+        message: "Camera access denied",
+      })
+    )
+    const select = new SelectFromPhotoLibrary(emptyProps)
+    await select.onPressNewPhoto()
+
+    const call = alert.mock.calls[0]
+    expect(call[0]).toEqual("Camera access denied")
+    expect(call[1]).toMatch(/enable/i)
+
+    const settingsButton = call[2][1]
+    settingsButton.onPress()
+    expect(Linking.openURL).toHaveBeenCalledWith(
+      NativeModules.ARCocoaConstantsModule.UIApplicationOpenSettingsURLString
+    )
+  })
+
+  it("shows an alert when saving a photo fails", async () => {
+    triggerMock.mockImplementationOnce(() =>
+      Promise.reject({
+        code: ARTakeCameraPhotoModule.errorCodes.saveFailed,
+        message: "Failed to save",
+        userInfo: {
+          NSUnderlyingError: {
+            code: 42,
+            message: "You have no hard disk",
+          },
+        },
+      })
+    )
+    const select = new SelectFromPhotoLibrary(emptyProps)
+    await select.onPressNewPhoto()
+    expect(alert).toHaveBeenCalledWith("Failed to save", "You have no hard disk (42)")
   })
 })